Fix type conversion in FormDeleteTodo to handle JSON unmarshaled float64
JSON unmarshaling returns numeric values as float64, so the id field needs
to be converted from float64 to int before comparison.

+++ src/routes/todos.remote.go
@@ -26,7 +26,7 @@ func FormCreateTodo(postData map[string]any) (any, error) {
 }
 
 func FormDeleteTodo(postData map[string]any) (any, error) {
-	id := postData["id"].(int)
+	id := int(postData["id"].(float64))
 	for i, todo := range todos {
 		if todo.ID == id {
 			todos = append(todos[:i], todos[i+1:]...)
